Handle multiple namespace files
To use multiple namespace files, the triple slash directive is added to the file requiring access to the namespace functions and properties.

Since it's not possible to require files in ES5 or the browser a module loader is needed. I added a few compiler options and installed the SystemJS loader.