Remove files from previous modules
The typescript compiler does not realize the files in the directory are not related. I attempted to fix this issue by only compiing the app.ts file and outputting only the app.js file. However, as soon as the lessons moved into the module and namespace topic, additional problems began to surface such as using the outFile option with the CommonJS module option.

To deal with this issue I'm simply going to work with the exercises as if the others never existed. Then when I push the changes to GitHub I can just refer to a particular commit to recall the lesson.