Loading Modules
In this lesson, I moved the geometric functions into separate files based on their focus, circle.ts and rectangle.ts. I moved the files into a new directory to simulate how a real project might be laid out.

In each of the module files the functions that will be imported must be exported. In the app.ts file the import statements are used to provide access to the functions. In the background, typescript is converting the import statements to require statements per the CommonJS standard.

To make this work in the browser requires a module loader. This tutorial uses SystemJS which is installed via npm and configured using the instructions on the documentation page. In short, the index.html file needs to add the dependency and include the configuration to import my app.js file.