Include unknown properties in the interface
Since it is possible that an object may contain properties that are not of interest or could not be predicted, an interface can account for those properties too. Use the '[]' around a placeholder property key with type 'string' (since most property keys are strings). The value can be assigned the 'any' type unless you already know a particular type to which it will be limited.

This prevents compilation errors when other properties exist on some objects, but not others, such as the hobbies array in the example.