Working with Generics
Using generics in typescript is a method of handling situations in which the type, perhaps unknown or limited, needs to be consistent throughout the component. Using a symbol representation, the type can be checked to confirm it matches when desired. For example, when  a function that takes in a string should return a string and when it takes a number it should return a number, using a generic will retain the advantages of typescript by enforcing type compliance. If the <any> type were used it would prevent compile checks for specific or expected types.